UTF-8 overhaul, part 3: libwdi + additional installer fixes
libwdi now saves the inf and provides parameters to the installer in UTF-16

+++ libwdi/logging.c
@@ -108,7 +108,7 @@ void pipe_wdi_log_v(enum wdi_log_level level,
 
 	if (truncated) {
 		WriteFile(logger_wr_handle, truncation_notice,
-			strlen(truncation_notice)+1, &junk, NULL);
+			(DWORD)strlen(truncation_notice)+1, &junk, NULL);
 		log_messages_pending++;
 		PostMessage(logger_dest, logger_msg, (WPARAM)level, 0);
 	}
